requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.8.30
charset-normalizer==3.4.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.16
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
idna==3.10
importlib-metadata==8.5.0
itypes==1.2.0
jinja2==3.1.4
Markdown==3.7
MarkupSafe==2.1.5
pillow==10.4.0
requests==2.32.3
sqlparse==0.5.2
typing-extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
zipp==3.20.2
